Data Scraping Prices
This project was the beginnings of an app that would help the travel
agency automatically detect suboptimal prices of its products on the
ski-express.com marketplace. That is a fancy way of saying: a tool to
help put the right price on the product compared to what else is being
sold.

input is an array of urls to read
these urls are pages on the ski-express site representing offers on
sale for a given trip
there can be an arbitrary number of urls

info about the output:
data written to CSV: Rang;Cible;Malin;MoinsCher;NbSejours
rang is the company (VTR) ranking on the list. best is 1
cible is the lowest price on the list
malin is 1 or 0 according to the whether or not the site says the
company product is malin or a smart buy
moins cher is 1 or 0 according to whether or not the site says the
company product is the cheapest on the list
nb sejours is the total number of products on the list

I think this code demonstrates reasonable understanding with
xpath in php
writing csv files

use by running "benchmark.php"
read file in "output" folder